% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{miss.imp}
\alias{miss.imp}
\title{imputate the missing data}
\usage{
miss.imp(data, method = "mean")
}
\arguments{
\item{data}{the matrix with missing data}

\item{method}{the method used to imputate the missing data}
}
\value{
a complete matrix with the missing data imputated
}
\description{
imputate the missing data
}
\examples{
\dontrun{
data <- matrix(c(1, NA, 3, 4, 5, 6), ncol = 2)
res <- miss.imp(data, method='reg')
}
}
