% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{miss.view}
\alias{miss.view}
\title{view the pattern of missing data}
\usage{
miss.view(data, type = "bar")
}
\arguments{
\item{data}{the matrix with missing data}

\item{type}{the form of the output}
}
\value{
the barplot or heatplot of the missing data
}
\description{
view the pattern of missing data
}
\examples{
\dontrun{
data <- matrix(c(1, NA, 3, 4, 5, 6), ncol = 2)
res <- miss.view(data, type='heat')
}
}
