% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwC_f}
\alias{rwC_f}
\title{A Gibbs sampler using Rcpp}
\usage{
rwC_f(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard error}

\item{x0}{the start point}

\item{N}{the number of circles}
}
\value{
a numeric matrix of results
}
\description{
A Gibbs sampler using Rcpp
}
