\name{StatComp20037-package}
\alias{StatComp20037-package}
\alias{StatComp20037}
\docType{package}
\title{
  The package is used to view and imputate the missing data.
}
\description{
  The package contains two function: miss.view (to view the missing data) and 
  miss.imp (to imputate the missing data).
}

\author{
Peng Chen

Maintainer: Peng Chen <chenpeng1@mail.ustc.edu.cn>
}
\keyword{ package }